+++ packages/opencensus-resource-util/test/test-detect-resource.ts
